Fix #2.
This commit isn't really required, but since the configuration used
in this repo also serves as an example, the following might confuse
people. Before this commit when building locally using mock data,
the latest tag would be used as a reqirect from the root URL.
Now the local build is redirected to instead.

* docs/poly.py (MOCK_DATA): Change type of the local revision to `BRANCH`.