Only copy files not ignored by git for local build (#31)
Before all files in the working directory would be copied to a temp dir for the local build.
This might include large directories that aren't directly needed and therefore specified in a `.gitignore` file. This commit tries to call git to list only files that aren't ignored by git.
If that fails the old method serves as a fall back. Otherwise the only the non-ignored files are copied.

* sphinx_polyversion/driver.py (Driver.build_local): Do not handle copying directly but call `VersionProvider.checkout_local`.

* sphinx_polyversion/vcs.py (VersionProvider.checkout_local): Define new abstract method.

* sphinx_polyversion/git.py (_get_unignored_files): Implement using `git ls-files --others --cached --exclude-standard`.

* sphinx_polyversion/git.py (Git.checkout_local): Implement as described above.

---------

Co-authored-by: real-yfprojects <62463991+real-yfprojects@users.noreply.github.com>